refactor: move the two security classes out of the vendored toolkit
HtmlSanitizer and RemoteAddress were filed under lib/Tools/, which is otherwise
a re-namespaced copy of daita/my-small-php-tools from the Nextcloud 22 era.
They are not part of it: both are recent, final, fully typed and documented, and
both are security controls -- an allowlist sanitizer over remote HTML on import,
and the guard that refuses outbound requests aimed at the local network.

They now sit beside PrivateKeyCipher and SecretHasher in lib/Security/, where
someone auditing the app's security controls will find them. Their tests move
with them.

This also makes the rest of lib/Tools/ a coherent thing to argue about:
27 files and 3,091 lines, all of it 2022-era vendored code, with nothing recent
mixed in to complicate the case for retiring it.

Note for anyone repeating this: composer.json sets classmap-authoritative, so a
moved class is not found until `composer dump-autoload` runs, and the failure
looks like the class was deleted.
