+++ tests/Security/HtmlSanitizerTest.php
@@ -7,9 +7,9 @@ declare(strict_types=1);
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-namespace OCA\Social\Tests\Tools;
+namespace OCA\Social\Tests\Security;
 
-use OCA\Social\Tools\HtmlSanitizer;
+use OCA\Social\Security\HtmlSanitizer;
 use PHPUnit\Framework\TestCase;
 
 class HtmlSanitizerTest extends TestCase {
